Refactor admin page
